Added array to pass to backend
