+++ src/components/food-selection/all-data.js
