+++ src/components/food-selection/product-information.js
